requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0
timm>=0.9.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
tqdm>=4.64.0
Pillow>=9.0.0
numpy>=1.21.0
pandas>=1.4.0